refactor: extract some helpers into separate module to avoid circle imports
